% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_to_plot_list}
\alias{add_to_plot_list}
\title{Add a plot to a plot_list}
\usage{
add_to_plot_list(plot_list, plot, filename = NULL)
}
\arguments{
\item{plot_list}{a list to hold the plots}

\item{plot}{a plot object}

\item{filename}{character, the filename to use to save the plot}
}
\value{
the new plot list
}
\description{
\code{add_to_plot_list} adds a plot and an optional file name
to the supplied list. If no file name is supplied an arbitrary
one is created of the form 'plot-' \code{index} '.png'.
}
\examples{
plot_list <- add_to_plot_list(plot_list, plot)
plot_list <- add_to_plot_list(plot_list, plot, filename = 'file1.png')

}
