% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLAST_to_OTU.R
\name{BLAST_to_OTU}
\alias{BLAST_to_OTU}
\title{Take a blast out put and count number of unique hits}
\usage{
BLAST_to_OTU(df, query = NA)
}
\arguments{
\item{df}{Dataframe with BLAST data. Should contain query name and output accesion ID.}

\item{query}{Vector of character to subset and merge by}
}
\value{
Dataframe with unique blast results and number of hits
}
\description{
Take a blast out put and count number of unique hits
}
