% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_all.r
\name{clear_all}
\alias{clear_all}
\title{Clears the workspace, console and shuts down all open graphics devices.}
\usage{
clear_all()
}
\description{
Clears the workspace, console and shuts down all open graphics devices.
}
