% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMatrix.R
\name{cluster_matrix}
\alias{cluster_matrix}
\title{hierarchical cluster matrix by rows or columns (or both)}
\usage{
cluster_matrix(
  mat,
  by_row = TRUE,
  by_col = FALSE,
  scale = TRUE,
  dist_method = "pearson",
  ...
)
}
\arguments{
\item{mat}{Matrix of counts to cluster}

\item{by_row}{logical - whether to cluster the rows of the matrix}

\item{by_col}{logical - whether to cluster the columns of the matrix}

\item{scale}{logical indicating whether to center and scale the data}

\item{dist_method}{character method for ceating the distance matrix}

\item{...}{other arguments passed to hclust}
}
\value{
reordered matrix based on clustering
}
\description{
\code{cluster_matrix} hierarchically clusters the supplied matrix by row or column
or both
}
\examples{
set.seed(1452)
cluster_matrix( matrix(rnorm(100), ncol = 5) )

cluster_matrix( matrix(rnorm(100), ncol = 5), by_row = FALSE, by_col = TRUE )

}
