% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColumnExpander.r
\name{column_expander}
\alias{column_expander}
\title{Column Expander}
\usage{
column_expander(data, name)
}
\arguments{
\item{data}{Any Data Set}

\item{name}{Name of the Column Variable whose value needs to be expanded}
}
\value{
Data set with additional columns with vules 0 and 1
}
\description{
Expands Values to Columns
}
\details{
Gives Data Set with levels / values as column variable with vales
0 and 1. 1's are the value where that value occurs in the variable rest all
are zero
}
\examples{
data(iris)
column_expander(iris,"Species")

}
\author{
Jitin Kapila
}

