% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense_blast.R
\name{condense_blast}
\alias{condense_blast}
\title{Take a blast out put and count_col number of unique hits}
\usage{
condense_blast(data, subset_values = NA, subset_col = 1, count_col = 2)
}
\arguments{
\item{data}{Data frame of the BLAST output}

\item{subset_values}{Subset results by character string in query name}

\item{count_col}{The countumn to count unique entries from. Can be countumn name or index}
}
\value{
Dataframe with unique blast results and number of hits
}
\description{
Take a blast out put and count_col number of unique hits
}
