% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_to_clipboard.R
\name{copy_to_clipboard}
\alias{copy_to_clipboard}
\title{Copy a dataframe in to the clipboard.}
\usage{
copy_to_clipboard(.dat, .sep = "\\t", .na = "")
}
\arguments{
\item{.dat}{\code{data.frame} to copy.}

\item{.sep}{Separator for the cells. To copy to excel use the default (\code{"\\t"}).}

\item{.na}{String to replace \code{NA} in the \code{data.frame}. The default is an empty string (\code{""}).}
}
\description{
Copy a dataframe in to the clipboard.
}
