% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyToClipboard.R
\name{CopyToClipboard}
\alias{CopyToClipboard}
\title{Copy Stuff to Clipboard}
\usage{
CopyToClipboard(data)
}
\arguments{
\item{R}{elements (tested only on vectors)}
}
\description{
Simple function to copy R elements to the clipboard for easy pasting
}
\examples{
CopyToClipboard(data)
}
\keyword{and,}
\keyword{copy,}
\keyword{paste}

