% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_unique.R
\name{count_unique}
\alias{count_unique}
\title{Takes a dataframe and count_col unique entries of a column
by grouping of id in another column.}
\usage{
count_unique(data, subset_col, count_col, subset_values = NA)
}
\arguments{
\item{data}{Dataframe.}

\item{subset_col}{Column to group counting by}

\item{count_col}{Column ID to count unique entries from.}

\item{subset}{Character vector with entries to subset by.}
}
\value{
Dataframe with number of unique entries in each subgroup
}
\description{
Takes a dataframe and count_col unique entries of a column
by grouping of id in another column.
}
