% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_pca.R
\name{data_to_pca}
\alias{data_to_pca}
\title{data_to_pca: From tidy data.table estimate PCAs for well-defined groups}
\usage{
data_to_pca(dt_to_pca, id1, id2, value, n_pca = 2, ..., scale = T,
  plot = F, verbose = T)
}
\arguments{
\item{dt_tmp:}{Tidy dataset}

\item{id1:}{vector of y (vertical) coordinates for the first line}

\item{id2:}{vector of x (horizontal) coordinates for the second line}

\item{value:}{vector of y (vertical) coordinates for the second line}

\item{n_pca:}{Number of PCAs to output}

\item{scale:}{Scaling option for the PCA analysis}

\item{plot:}{plot "time series" of PCA}

\item{verbose:}{non binding for now}
}
\value{
list of datasets (and plot) of loadings and time series
}
\description{
data_to_pca: From tidy data.table estimate PCAs for well-defined groups
}
\note{
Useful for some time series representations
}
\examples{
  data_to_pca(dt_ind, date_y, ind, n_pca = 5)
}
