% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_order_change.r
\name{date_order_change}
\alias{date_order_change}
\title{Changes date format orders}
\usage{
date_order_change(x, from_formats, to_formats)
}
\arguments{
\item{x}{A character or numeric vector of dates.}

\item{from_formats}{A character vector of formats. Should be the same length as \code{to_formats}.}

\item{to_formats}{A character vector of formats. Should be the same length as \code{from_formats}.}
}
\value{
A character vector with dates changed into \code{to_format}.
}
\description{
Changes date format orders
}
\examples{
dates <- c("10/23/2022", "01/01/2018", "10/2022", "01/2018", "2019/05", "2025")
date_order_change(dates,
                  from_formats = c("\%m/\%d/\%Y", "\%Y-\%m-\%d", "\%Y/\%m", "\%m/\%Y", "\%Y"),
                  to_formats = c("\%d/\%m/\%Y", "\%d/\%m/\%Y", "\%m/\%Y", "\%m/\%Y", "\%Y"))
}
