% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datem_convert.R
\name{datem_conv}
\alias{datem_conv}
\title{datem_conv: convert date_ym and datem formats}
\usage{
datem_conv(date, fulldate = F)
}
\arguments{
\item{date:}{a monthly date either in yyyymm format or in monthly statar format}

\item{fulldate:}{boolean}
}
\value{
A date
}
\description{
datem_conv: convert date_ym and datem formats
}
\note{
Useful to juggle with date types in monthly panel
}
\examples{
  datem_conv(199901)
}

