% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualplot.R
\name{dualplot}
\alias{dualplot}
\title{dualplt: dualplot function from Peter Ellis (github.com/ellisp) for time series plots on double axes}
\usage{
dualplot(x1, y1, y2, x2 = x1, col = c("#C54E6D", "#009380"), lwd = c(1,
  1), colgrid = NULL, mar = c(3, 6, 3, 6) + 0.1,
  ylab1 = paste(substitute(y1), collapse = ""),
  ylab2 = paste(substitute(y2), collapse = ""), nxbreaks = 5,
  yleg1 = paste(gsub("\\n$", "", ylab1), "(left axis)"),
  yleg2 = paste(ylab2, "(right axis)"), ylim1 = NULL, ylim2 = NULL,
  ylim.ref = NULL, xlab = "", main = NULL, legx = "topleft",
  legy = NULL, silent = FALSE, bty = "n", ...)
}
\arguments{
\item{x1:}{vector of x (horizontal) coordinates for the first line}

\item{y1:}{vector of y (vertical) coordinates for the first line}

\item{x2:}{vector of x (horizontal) coordinates for the second line}

\item{y2:}{vector of y (vertical) coordinates for the second line}

\item{col:}{vector of colors}
}
\value{
ggplot object
}
\description{
dualplt: dualplot function from Peter Ellis (github.com/ellisp) for time series plots on double axes
}
\note{
Useful for some time series representations
}
\examples{
  dualplot(1:100, runif(100), 10*runif(100))
}
