% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emailme.R
\name{emailme}
\alias{emailme}
\title{Email Me}
\usage{
emailme(out = NULL, who = "erikl@mit.edu",
  outgoing_server = "outgoing.mit.edu")
}
\arguments{
\item{out:}{R output}

\item{who:}{who is sending / receiving}

\item{outgoing_server:}{specify smtp}
}
\description{
Email Me
}
\note{
print the current environment. package is meant to be used or database.
  Uses pretty print to get an idea of how crowded is the memory on the system.
}
