% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_case_when.r
\name{fct_case_when}
\alias{fct_case_when}
\title{Adds factor levels to a case_when}
\usage{
fct_case_when(...)
}
\arguments{
\item{...}{\code{dplyr::case_when()} code.}
}
\value{
Output from \code{case_when()} with factor levels added. The factor order is the same as specified in the syntax.
}
\description{
Adds factor levels to a case_when
}
\examples{
library(dplyr)

df <- mutate(starwars,
             height_cat = fct_case_when(
               height <= 150 ~ "<= 150",
               height > 150 ~ "> 150"
             ))

levels(df$height_cat)


}
