% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pkgs.r
\name{find_pkgs}
\alias{find_pkgs}
\title{Get packages used in a file or a directory.}
\usage{
find_pkgs(file = NULL, dir = NULL, details = FALSE, simplify = TRUE)
}
\arguments{
\item{file}{A file path to search for packages.}

\item{dir}{A directory path to search for packages. If \code{file} is already specified this is ignored. See details.}

\item{details}{A logical value. Is a detailed description of how packages are invoked needed. The default is \code{FALSE}.}

\item{simplify}{A logical value indicating if the \code{dir} name should be removed from file names in details. This value is ignored if a \code{file} is  specified or if \code{details = FALSE}.}
}
\value{
A character vector of packages in the file if \code{details = FASLE}. If \code{details = TRUE}, a list with 2 elements. A character vector of packages in \code{pkgs} and a details dataframe in \code{details_df}.
}
\description{
Get packages used in a file or a directory.
}
\details{
When scanning a directory only files with the extension \code{.r} or \code{.R} searched.
}
