% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_number.r
\name{format_number}
\alias{format_number}
\title{Formats a given numeric vector with the specified number of decimals}
\usage{
format_number(x, .output_digits)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{.output_digits}{Number of output digits.}
}
\value{
A formatted character vector of the values.
}
\description{
Formats a given numeric vector with the specified number of decimals
}
