% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_case_when_code.r
\name{get_case_when_code}
\alias{get_case_when_code}
\title{Outputs the \code{case_when} code given classes in a table format}
\usage{
get_case_when_code(
  classes_dat,
  response_col,
  output_file,
  dplyr_code = TRUE,
  dataset_name = NULL
)
}
\arguments{
\item{classes_dat}{Dataset with the conditionas and the responses.}

\item{response_col}{The name of the response column}

\item{output_file}{A file path for the output file.}

\item{dplyr_code}{Should the code be compatible with \code{dplyr} programming. The default is \code{TRUE}. If \code{FALSE} the dataset name will be appended to all variables.}

\item{dataset_name}{Dataset name. Required if \code{dplyr = FALSE}.}
}
\description{
Outputs the \code{case_when} code given classes in a table format
}
\details{
A case will be defined for each row with the conditions joined by \code{&}. The column names are considered as the variable names. There can be OR (\code{|}) statements inside cells.
}
