% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gg_legend.R
\name{get_gg_legend}
\alias{get_gg_legend}
\title{get_gg_legend}
\usage{
get_gg_legend(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{a ggplot object}
}
\value{
the legend as a grob
}
\description{
\code{get_gg_legend} returns the legend of a ggplot object as a grob
}
\examples{
plot_legend <- get_gg_legend(ggplot_obj)

}
