% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_perc_txt.r
\name{get_n_perc_txt}
\alias{get_n_perc_txt}
\title{Get the n(\%) text, given n and the proportion separately.}
\usage{
get_n_perc_txt(n, prop, .output_digits = c(0, 1), .with_percentage = TRUE)
}
\arguments{
\item{n}{A vector or a single value of counts.}

\item{prop}{A vector or a single value of proportions}

\item{.output_digits}{Number of decimal points in the output percentages. If a numeric vector with more than 2 elements provided,
the first number will be taken as the number of digits for n (default is 0) and the second number as the digits for the percentage (default is 1).}

\item{.with_percentage}{A logical indicating if the percentage mark (\%) should be shown. The default is \code{TRUE}.}
}
\value{
A vector or a single string as n(\%).
}
\description{
Get the n(\%) text, given n and the proportion separately.
}
