% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_num_3_txt.r
\name{get_num1_num2_num3_txt}
\alias{get_num1_num2_num3_txt}
\title{Returns a string as num1 (num2, num3) given num1 and num3 separately}
\usage{
get_num1_num2_num3_txt(num1, num2, num3, .output_digits = 2)
}
\arguments{
\item{num1}{A vector or a single value of num1 (numerical) value(s).}

\item{num2}{A vector or a single value of num2 (numerical) value(s).}

\item{num3}{A vector or a single value of num3 (numerical) value(s).}

\item{.output_digits}{Number of decimal points in the output percentages. Accepts vectors with 2 elements. If a vector with more than 2 elements is provided,
the first 2 items are selected with a warning. The default is set to 2.}
}
\value{
A string pasted as num1 (num2, num3).
}
\description{
Returns a string as num1 (num2, num3) given num1 and num3 separately
}
