% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_num_summary.r
\name{get_num_summary}
\alias{get_num_summary}
\title{Get the numerical summary by multiple categorical by variables}
\usage{
get_num_summary(.data, .num_var, ..., .na.rm = FALSE, .output_digits = 2)
}
\arguments{
\item{.data}{A data frame or a data frame extension (e.g. a tibble).}

\item{.num_var}{Primary (numerical) variable of interest.}

\item{...}{Additional variables to count by. This can be left blank.}

\item{.na.rm}{A logical value indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{.output_digits}{Number of decimal points in the output for the mean and standard deviation. The default is set to 2.}
}
\value{
A tibble with the by variables (if specified in \code{...}) and 3 columns. Number of values (\code{N}), Number of non-missing values (\code{n}), Mean-SD (\code{mean_sd}), Median-Q1-Q3 (\code{median_q1_q3}) and Min-Max (\code{min_max}).
The number of rows depends on the combinations of the by variables, if specified. Will have one row if by variables are not specified.
}
\description{
Get the numerical summary by multiple categorical by variables
}
\examples{
library(dplyr)

starwars \%>\%
get_num_summary(.num_var = height,
                species, gender,
                .na.rm = TRUE,
                .output_digits = 1)
}
