% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_theme.R
\name{gg_bw}
\alias{gg_bw}
\title{Mostly Black and White theme}
\usage{
gg_bw(base_size = 10)
}
\arguments{
\item{base_size:}{position}
}
\value{
A theme for ggplot2
}
\description{
Mostly Black and White theme
}
\note{
print the current environment. package is meant to be used or database.
  Uses pretty print to get an idea of how crowded is the memory on the system.
}
\examples{
data(mtcars); library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) + geom_point()
p + gg_bw()
}
