% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab_load.R
\name{gitlab_load}
\alias{gitlab_load}
\title{Load from gitlab}
\usage{
gitlab_load(package = "wrdsr", repo = "gitlab.com",
  username = "Loualiche", package_dir = "~/.R/tmp_dir/")
}
\arguments{
\item{package:}{gitlab package}

\item{repo:}{gitlab package address}

\item{package_dir:}{dir of where to load the gitlab R package}
}
\description{
Load from gitlab
}
\note{
print the current environment. package is meant to be used or database.
  Uses pretty print to get an idea of how crowded is the memory on the system.
}
