% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfoContineous.r
\name{infogen_continuous}
\alias{infogen_continuous}
\title{Continuous Information}
\usage{
infogen_continuous(data)
}
\arguments{
\item{data}{Any Data Set}
}
\value{
Information about continuous Variables
}
\description{
Gives all information about continuous variable in data.
}
\details{
Gives Max., Min., Mean. Median, Std. Div. Counts, Unique Counts,
NA's, And Quantile
}
\examples{
data(iris)
infogen_continuous(iris)

data(mtcars)
info <- infogen_continuous(mtcars)
print(info)
}
\author{
Jitin Kapila
}

