% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_missing_packages.r
\name{install_missing_pkgs}
\alias{install_missing_pkgs}
\title{Installs specified packages from CRAN if missing.}
\usage{
install_missing_pkgs(cran_pkgs, repos = "https://cloud.r-project.org")
}
\arguments{
\item{cran_pkgs}{A character vector with package names as elements.}

\item{repos}{A character vector of the base URL(s) of the repositories to use. The default is CRAN ("https://cloud.r-project.org")}
}
\description{
Installs specified packages from CRAN if missing.
}
\examples{
install_missing_pkgs(cran_pkgs = c("tidyverse"))
}
