% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_objects.R
\name{ls_objects}
\alias{ls_objects}
\title{List objects loaded in current environment}
\usage{
ls_objects(pos = 1, order.by = "Size", pattern = "", decreasing = TRUE,
  head = FALSE, n = 5, database = TRUE)
}
\arguments{
\item{pos:}{position}

\item{order.by:}{ordering}

\item{pattern:}{which objects to print}

\item{decreasing:}{order of print}

\item{head:}{how to print it}

\item{n:}{number of objects to print}

\item{database:}{only print database objects (data.table and data.frame)}
}
\value{
A string.
}
\description{
List objects loaded in current environment
}
\note{
print the current environment. package is meant to be used or database.
  Uses pretty print to get an idea of how crowded is the memory on the system.
}
\examples{
ls_objects()
}

