% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapValuesMany.R
\name{MapValuesMany}
\alias{MapValuesMany}
\title{Map a number of multiple values to single values (code by Mark Riseley)}
\usage{
MapValuesMany(x, from, to)
}
\arguments{
\item{x, }{from, to}
}
\description{
to is a single target recoding value for each list main element in from,
e.g. from = list(1:10, 11:20), to = c("1-10"; "11:20")
}
\examples{
MapValuesMany(x,from,to)
}
\keyword{map}
\keyword{recode,}
\keyword{vectors,}

