% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{open_graphics_device}
\alias{open_graphics_device}
\title{Open a graphics device based on a file name}
\usage{
open_graphics_device(filename = "plot.pdf", ...)
}
\arguments{
\item{filename}{character, the filename to use}

\item{...}{Arguments passed on to the graphics device}
}
\value{
TRUE invisibly
}
\description{
\code{open_graphics_device} takes a file name and opens a graphics
device based on the file suffix. Defaults to pdf if no match is found.
}
\examples{
open_graphics_device(filename = 'plot.eps')
open_graphics_device(filename = 'plot.eps', width = 6, height = 5)
open_graphics_device(filename = 'plot.png', width = 640, height = 480)

}
