% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenAllRFiles.R
\name{OpenAllRFiles}
\alias{OpenAllRFiles}
\title{Open All R Files in Current Directory for Editing}
\usage{
OpenAllRFiles()
}
\description{
Simple shortcut function to open all R files in current working directory for editing in RStudio
}
\examples{
OpenAllRFiles()
}
\keyword{convenience,}
\keyword{editing}
\keyword{file,}
\keyword{wrapper,}

