% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{output_plot}
\alias{output_plot}
\title{Output a plot from a plot list}
\usage{
output_plot(plot_entry, ...)
}
\arguments{
\item{plot_entry}{list containing filename and plot}

\item{...}{Arguments passed on to the graphics device}
}
\value{
the results of \code{dev.off()} invisibly
}
\description{
\code{output_plot} take a plot_entry object and saves the plot using
the correct graphics device. Currently the only extra parameters that
can be passed through to the graphics device are width, height and bg
as these are the only ones that are commmon between all 4 devices,
}
\examples{
output_plot(plot_entry)
output_plot(plot_entry, width = 6, height = 7)

}
