% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_lineage_from_accesion.R
\name{tax_lineage_from_accesion}
\alias{tax_lineage_from_accesion}
\title{Get full taxonomic lineage from accession nucleotide ID}
\usage{
tax_lineage_from_accesion(accession, custom_tax = NA)
}
\arguments{
\item{accession}{Accesion ID for the nucleotide database}

\item{custom_tax}{Character vector with taxonomic levels to return}
}
\value{
Dataframe with taxonomic info for each accession
}
\description{
Get full taxonomic lineage from accession nucleotide ID
}
