% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMatrix.R
\name{cluster}
\alias{cluster}
\title{hierarchical clustering}
\usage{
cluster(mat, scale = TRUE, dist_method = "pearson", ...)
}
\arguments{
\item{mat}{Matrix of counts to cluster}

\item{scale}{logical indicating whether to center and scale the data}

\item{dist_method}{character method for ceating the distance matrix}

\item{...}{other arguments passed to hclust}
}
\value{
matrix with columns reordered based on clustering
}
\description{
\code{cluster} hierarchically clusters the supplied matrix by column and
returns the reordered matrix
}
\examples{
set.seed(1452)
cluster( matrix(rnorm(100), ncol = 5) )

}
