% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense_blast.R
\name{condense_blast}
\alias{condense_blast}
\title{Take a blast out put and count number of unique hits}
\usage{
condense_blast(df, query_subset = NA)
}
\arguments{
\item{df}{Data frame of the BLAST output}

\item{query_subset}{Subset results by character string in query name}
}
\value{
Dataframe with unique blast results and number of hits
}
\description{
Take a blast out put and count number of unique hits
}
