% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_subtrees.R
\name{extract_subtrees}
\alias{extract_subtrees}
\title{extract_subtrees}
\usage{
extract_subtrees(graph, nodes)
}
\arguments{
\item{graph}{a graphNEL object}

\item{nodes}{a character vector of node labels to produce subtrees for}
}
\value{
a list of graphNEL objects
}
\description{
\code{extract_subtrees} extracts unconnected subtrees given a graph and a list of nodes
}
\details{
This function requires the \code{\link[topGO]} and \code{\link[graph]} packages to be installed.
}
\examples{
listOfSubTrees <- extract_subtrees(graph, nodes)

}
